+++ frontend/src/containers/issue/components/SearchFilter.js
@@ -2,6 +2,7 @@ import styled from 'styled-components';
 import React, { useState } from 'react';
 import dropdownCaret from '../../../assets/styles/caret';
 import { svgX } from '../../../assets/svgPath';
+import Modal from '../../../components/Modal'
 
 const SearchFilterWrapper = styled.div`
     position: relative;
@@ -21,20 +22,6 @@ const SearchFilterWrapper = styled.div`
     *:focus{
       outline:none;
     }
-`;
-
-const FilterItemList = styled.div`
-width: 300px;
-height: auto;
-position: absolute;
-max-height: 480px;
-margin: 8px 0 16px;
-font-size: 12px;
-top: 78%;
-left: 0%;
-border-color: #eaecef;
-border-radius: 6px;
-border: 1px solid #eaecef;
 `;
 
 const FirstEl = styled.div`
@@ -78,23 +65,6 @@ const filterItems = (closeFunc, items) => (
     {items.map((item) => (<FilterItem>{item}</FilterItem>))}
   </>
 );
-
-const Modal = ({
-  className, onClose, visible, children,
-}) => {
-  const onMaskClick = (e) => {
-    if (e.target === e.currentTarget) {
-      onClose(e);
-    }
-  };
-  return (
-    <>
-      <FilterItemList className={className} onClick={onMaskClick} tabIndex="-1" visible={visible}>
-        {children}
-      </FilterItemList>
-    </>
-  );
-};
 
 const SearchFilter = () => {
   const [modalVisible, setModalVisible] = useState(false);
